Bug fix 1-2-3-gnn class.

+++ examples/1-2-3-gnn-class.py
@@ -103,8 +103,8 @@ class Net(torch.nn.Module):
         # final_classification layers
         setattr(self, 'fc0', torch.nn.Linear(args.max_k*args.emb_dim, args.emb_dim))
         for l in range(1, args.num_linear_layers-1):
-            setattr(self, 'fc{}'.format(l), torch.nn.Linear(args.emb_dim, args.emb_dim/2))
-        setattr(self, 'fc{}'.format(args.num_linear_layers-1), torch.nn.Linear(max(4,args.emb_dim/(2^(args.max_k-2))), dataset.num_classes))
+            setattr(self, 'fc{}'.format(l), torch.nn.Linear(args.emb_dim, int(int(args.emb_dim)/2)))
+        setattr(self, 'fc{}'.format(args.num_linear_layers-1), torch.nn.Linear(max(4,int(int(args.emb_dim)/(2^(args.max_k-2)))), dataset.num_classes))
 
     def reset_parameters(self):
         for (name, module) in self._modules.items():
